+++ logs/hostinfo.go
@@ -1,4 +1,4 @@
-package processor
+package logs
 
 import (
 	"github.com/cockroachdb/errors"